corrigi algums problemas e agora o texto vai para o centro da tela indepente do tamanho do campo

+++ core/src/com/mygdx/draw/TextDraw.java
@@ -7,8 +7,8 @@ public class TextDraw {
     public static void draw(SpriteBatch sprite ,Text text) {
     	BitmapFont font  = text.getFont();
     	String textString = text.getTextString();
-    	int coordinateX = text.getTextPosition().getCoordinateX();
-    	int coordinateY = text.getTextPosition().getCoordinateY();
+    	int coordinateX = (int) text.getTextPosition().getCoordinateX();
+    	int coordinateY = (int) text.getTextPosition().getCoordinateY();
     	
     	font.draw(sprite, textString, coordinateX, coordinateY);
